move useSharePlaylists()
