+++ projects/graphql-client/src/index.ts
@@ -11,3 +11,4 @@ export * from "./remote/queries/useViewer"
 export * from "./remote/queries/useSongSearch"
 export * from "./remote/queries/useShares"
 export * from "./remote/queries/useShareUsers"
+export * from "./remote/queries/useSharePlaylists"
